bluetooth-test/requirements.txt
RPi.GPIO
dbus-python
PyGObject
